fix: updated slots and docs
