+++ src/components/reusable/errorBlock/errorBlock.mdx
@@ -1,0 +1,41 @@
+import { Meta, Story, Canvas, Controls } from '@storybook/blocks';
+import * as ErrorStories from './errorBlock.stories.js';
+import alertImg from './redfords/Alert_Redford.svg';
+
+<Meta of={ErrorStories} />
+
+# Error
+Error block
+
+## Usage
+
+### JS Import
+```js
+import '@kyndryl-design-system/shidoka-applications/components/reusable/errorBlock';
+```
+
+### HTML Tag
+```html
+<kyn-error-block></kyn-error-block>
+```
+
+## Importing the error svg
+The following varients of error images/**Redford's** are available in the ```src/components/reusable/errorBlock/redfords```.
+- Alert Redford 
+- Not found Redford
+- Stop Redford
+- Timeout Redford
+
+Import the required svg as shown below:
+```js
+import timeoutImg from '@kyndryl-design-system/shidoka-applications/components/reusable/errorBlock/redfords/Timeout_Redford.svg';
+```
+
+## Primary Example
+
+<Canvas>
+  <Story of={ErrorStories.ErrorBlock} />
+</Canvas>
+
+### Properties Reference
+<Controls of={ErrorStories.ErrorBlock} />
