+++ src/components/reusable/errorBlock/index.ts
@@ -1,0 +1,1 @@
+export { ErrorBlock } from './errorBlock';
