--- src/components/reusable/errorComponent/index.ts
@@ -1,1 +1,0 @@
-export { ErrorComponent } from './errorComponent';
